ENH running tests
